% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_scrambler.R
\name{image_scrambler}
\alias{image_scrambler}
\title{An Image Scrambling Function}
\usage{
image_scrambler(path, chunks = 1, sep = "_", filetype = "tif")
}
\arguments{
\item{path}{What is the path to the folder containing the images?}

\item{chunks}{How many blocks of information should the image names be broken into as separated by "_"? Defaults to 1.}

\item{sep}{What string so you want to divide the file name with into various chunks}

\item{filetype}{What filetype are you scrambling}
}
\description{
This function copies the information from tif image files, renames the images with a scramble key, then saves two csv files, one with the scramble key and one with just the scramble numbeers.
}
\keyword{image}
\keyword{image_scrambler()}
\keyword{scrambler}
