% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_scrambler.R
\name{image_unscrambler}
\alias{image_unscrambler}
\title{An Image Unscrambling Function.}
\usage{
image_unscrambler(path)
}
\arguments{
\item{path}{What is the path to the folder containing the csv files?}
}
\description{
This function uses the key and scramble csv files generated by the image scrambler and recombines them into a functional csv file for analysis
}
\keyword{image}
\keyword{image_unscrambler()}
\keyword{unscrambler}
