% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_functions.R
\name{parse}
\alias{parse}
\title{An ITCN Results parser}
\usage{
parse(path, scramble = TRUE)
}
\arguments{
\item{path}{What is the path to the folder containing the results file?}
}
\description{
This function used the results txt file produced by the ImageJ plugin ITCN and parses it to extract the cell counts and area measurements.
}
\keyword{FIJI}
\keyword{ITCN}
\keyword{ImageJ}
\keyword{Results}
\keyword{parse}
\keyword{parse()}
